Update: move byte array utils
