+++ shared/util/src/main/kotlin/company/thewebhook/util/ByteArray.kt
@@ -1,0 +1,15 @@
+package company.thewebhook.util
+
+import java.nio.ByteBuffer
+
+fun Long.toByteArray(): ByteArray {
+    val buffer = ByteBuffer.allocate(Long.SIZE_BYTES)
+    buffer.putLong(this)
+    return buffer.array()
+}
+
+fun Int.toByteArray(): ByteArray {
+    val buffer = ByteBuffer.allocate(Int.SIZE_BYTES)
+    buffer.putInt(this)
+    return buffer.array()
+}
